Add PHP bindings for new features and improve code quality
This commit exposes the C implementations from previous commits as PHP
functions and improves overall code quality through phases 1-3 of the
quality improvement plan.

New PHP Functions:

Mouse support:
- tui_mouse_enable(mode): Enable mouse tracking
- tui_mouse_disable(): Disable mouse tracking
- tui_mouse_get_mode(): Get current mouse mode
- Constants: TUI_MOUSE_MODE_OFF/CLICK/BUTTON/ALL

Bracketed paste:
- tui_bracketed_paste_enable(): Enable bracketed paste mode
- tui_bracketed_paste_disable(): Disable bracketed paste mode
- tui_bracketed_paste_is_enabled(): Check if enabled

Clipboard (OSC 52):
- tui_clipboard_copy(text, target): Copy to clipboard
- tui_clipboard_request(target): Request clipboard contents
- tui_clipboard_clear(target): Clear clipboard

Input history:
- tui_history_create(maxEntries): Create history resource
- tui_history_destroy(history): Destroy history
- tui_history_add(history, entry): Add entry
- tui_history_prev/next(history): Navigate history
- tui_history_reset(history): Reset navigation position
- tui_history_save_temp/get_temp(history): Temp input storage

Property declarations:
- Box: tabIndex, focusGroup, autoFocus, focusTrap
- Text: hyperlink

Code quality improvements:
- Add -Wconversion flag to catch implicit narrowing
- Fix type conversion warnings with explicit casts
- Add debug assertion macros (src/debug.h)
- Add security documentation (docs/security.md)
- Add static analysis script (scripts/static-analysis.sh)
- Add Valgrind test script (scripts/test-valgrind.sh)
- Add 6 new tests for error paths and edge cases

+++ src/drawing/canvas.c
@@ -360,7 +360,7 @@ char** tui_canvas_render(tui_canvas *canvas, int *line_count)
                 if (tui_canvas_get(canvas, base_x + 1, base_y + 1)) bits |= 1;
 
                 const char *block = block_chars[bits];
-                int len = strlen(block);
+                size_t len = strlen(block);
                 memcpy(p, block, len);
                 p += len;
